vt: refactor frame repeats/field handling
Previously, field handling and repeating were dealt with in the render
thread.  This is cumbersome in the render thread and incompatable with
future refactoring of the GL rendering.

 - VideoData now has the concept of what field is current.
 - calling advance on the VideoTransport will move internal state to
   the correct next frame/field.  getFrame() may return a pointer
   to the same VideoData if advance() has not moved to another frame
   due to repeats.

The one possible downside to this is that we loose coupling between the
GL and the frame periods.  However, the current code doesn't have this
anyway.