fq: fix race causing frames to be inserted into wrong queue
If the videotransport changed direction while the frame queue was
re-populating the past/future lists, a frame from the past could be
appended inserted in the future list, and future frames could be
inserted into the past list.

Updates to direction are atomic, and guaranteed not to conflict with
the atomic read of direction, used to determine which lists to update.

Unforutnately the routine updating the list erroneously used direction
to work out the frame number offset, rather than infering the direction
from which list is being used.

Thanks to Stuart.Grace@bbc.co.uk for identifying the module causing the
fault.