vt: remove caching used frame queue
It is desirable for the frame queue not to be acting as an allocator.
This simplistic approach gets in the way of more complicated behaviours
that need to hold on to data for longer than a frame period.
Additionally, it complicates implementing more exotic allocators.
If a caching allocator is required, it should be implemented outside
of the frame queue.