GL: fix rendering using parameters from uploaded frame/field
Due to using two uploaders and renderers, the videoData obtained and
uploaded in the first half of the loop pertains to the current refresh.
But the rendering is conducted in the subsequent frame period.

Reorder the code so that the VideoData uploaded is still in scope
for the rendering (rather than rendering using the parameters of the
next frame).

This fixes the incorrect field ordering when de-interlacing.