vt: don't display each frame twice when deinterlacer disabled
When decoding interlaced material and deinterlacing, single stepping
advances to the next frame.  However, previously, with the deinterlacer
disabled, single stepping would show the same frame twice.  The intended
behaviour of single stepping is to always advance to the next frame/field.

A method is introduced to inform the VideoTransport of the state of
deinterlacing.  The behaviour is modified using:
  VideoTransport::setSteppingIgnoreInterlace()