QT: Hack to work around issues with QT, X11/XCB and GLX stalling
On some systems, when the GL rendering is done in a separate thread
using QT, the swapBuffers() call blocks, until the something kicks the
X event loop.  This only appears to happen with QT, a standalone GLX
program with multiple threads does not have this behaviour.

Work around it by opening a new connection to the X server and obtaining
a dedicated GL context.  All GL/GLX commands are then performed over
this display connection, avoiding the blocking issue.