build: replace qmake with autoconf/automake
QMake has an irritating limitation when it comes to configuring
projects, one must effectively hand edit the .pro file for anything
complicated.  Especially when libraries are in non-standard locations.

Autoconf allows this to be autodetected or overridden on a per build
directory basis.  Ie, it is far simpler to have a single tree produce
multiple different targets (eg Linux and Win32 binaries).

Currently two outputs are build:
 - libqttv.so/a: a library that encapsulates the QT widget and rendering
                 functions, but does not include the framequeues or IO.
 - videoplayer:  a binary that provides the whole videoplayer, linking
                 against libqttv

At the moment, the split is rather arbitrary, certain IO functions could
move to the library if required for instance.  It should also be noted
that the current GL renderer is almost entirely free from QT, so it
could be linked against non-QT applications too with only minor
modifications.

THe old videoplayer.pro file still exists, it might start to bitrot and
is only intended to build videoplayer itself.

Modifications by davidf:
 - rearranged source tree (to avoid splitting lib and app)
 - simplified (perhaps) generation of version.c
 - move all automake code to the toplevel (flat) Makefile.am
 - enabled silent rules by default
 - added host detection for platform specific conditional compilation
 - added enable/disable feature options