vt: allow customization of the read ahead and LRU list lengths
Adds two new options:
  --vt.readahead=N, Where N is the number of frames the frame queue can
                    hold ahead of the current display frame.
  --vt.lrucache=M,  Where M is the number of frames that have been
                    displayed to cache before releasing them.

For readahead, the videoplayer does not attempt to load all N frames
before starting playback.  It only limits the distance the reader
threads can get in front of the current display frame.

When adjusting the lru cache size, if playing in one direction only,
at a single speed, this can be very small.  The main use is to avoid
re-reading frames when changing direction, single stepping forwards or
backwards, or changing speed.