stats: use a thread id rather than object address to differentiate stat sections
Previously, when multiple classes used the same section name in the same
thread, they were assigned separate sections based on the object
address.  Eg, having a playlist with 10 items would create 10 yuvReader
entries.  Now, a single entry is created, and as a bonus, the thread id
can be used to match yuvReader to the frameQueue (same thread).